+++ src/core/ocr/base.ts
@@ -1,18 +1,7 @@
 import { chunk, Point, unique } from '@/common';
 import { createScheduler, createWorker } from 'tesseract.js';
-import {
-  BreachProtocolRawData,
-  COLS,
-  cross,
-  generateSquareMap,
-  HexNumber,
-  HEX_NUMBERS,
-  ROWS,
-} from '../common';
+import { HexNumber, HEX_NUMBERS } from '../common';
 import { FragmentId } from '../ocr';
-import { BreachProtocolBufferSizeFragmentResult } from './buffer-size';
-import { BreachProtocolDaemonsFragmentResult } from './daemons';
-import { BreachProtocolGridFragmentResult } from './grid';
 import { ImageContainer } from './image-container';
 
 export interface BreachProtocolFragmentBoundingBox {
@@ -186,43 +175,4 @@ export abstract class BreachProtocolOCRFragment<
   }
 
   static scheduler: Tesseract.Scheduler;
-}
-
-export class BreachProtocolRecognitionResult<C> {
-  readonly positionSquareMap = this.getPositionSquareMap();
-
-  readonly rawData = this.toRawData();
-
-  constructor(
-    public readonly grid: BreachProtocolGridFragmentResult<C>,
-    public readonly daemons: BreachProtocolDaemonsFragmentResult<C>,
-    public readonly bufferSize: BreachProtocolBufferSizeFragmentResult<C>
-  ) {}
-
-  toRawData(): BreachProtocolRawData {
-    return {
-      bufferSize: this.bufferSize.rawData,
-      daemons: this.daemons.rawData,
-      grid: this.grid.rawData,
-    };
-  }
-
-  private getPositionSquareMap() {
-    const { top, left } = this.grid.boundingBox;
-    const lines = this.grid.source.hocr
-      .split('\n')
-      .filter((l) => l.includes('ocrx_word'));
-    const size = Math.sqrt(lines.length);
-    const squares = cross(ROWS.slice(0, size), COLS.slice(0, size));
-    const bounds = lines
-      .map((l) => l.match(/(?<=bbox )(\d|\s)*/g)[0])
-      .map((l) => l.split(/\s/).map(Number));
-
-    return generateSquareMap(squares, (s, i) => {
-      const [x, y] = bounds[i];
-
-      // this is top left corner of each square
-      return new Point(x + left, y + top);
-    });
-  }
 }
